skills/group_meeting_recorder/requirements.txt
mss>=9.0.0
PyGetWindow>=0.0.9
Pillow>=10.0.0
numpy>=1.26.0
